Support swapping ref/alt alleles
