+++ test-data/liftover/picard/failed.mismatchRefSeq.testLiftoverBiallelicIndels.vcf
@@ -7,6 +7,4 @@
 ##INFO=<ID=AttemptedLocus,Number=1,Type=String,Description="The locus of the variant in the TARGET prior to failing due to mismatching alleles.">
 #CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	Sample1
 chr1	61	.	GT	G	724.43	ReverseComplementedIndel	.	GT	0/1
-chr1	72	.	T	A	100	MismatchedRefAllele	AttemptedLocus=chr1:469-469	GT	0/1
-chr1	72	.	T	A	100	MismatchedRefAllele	AttemptedLocus=chr1:469-469	GT	./.
 chr1	72	.	TT	A	100	ReverseComplementedIndel	.	GT	1/.
