+++ test-data/liftover/picard/testLiftoverSwapRefAltVariants.vcf
